+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## NEXT
+* Added an option to set the debounce timeout value which applies to all indexes by default, but can be customized for each index with a filter:
+
+Dynamic filter name: `algolia_autocomplete_debounce_{$index_name}_{$index_type}`
+
+Where `$index_name` is defined by the Index name prefix set on the WP Search with Algolia settings page and `$index_type` is the type of index.
+
+Assuming `wp_` is the Index name prefix, the debounce timeout filters would be:
+```
+algolia_autocomplete_debounce_wp_searchable_posts
+algolia_autocomplete_debounce_wp_post
+algolia_autocomplete_debounce_wp_page
+algolia_autocomplete_debounce_wp_my_custom_post_type
+algolia_autocomplete_debounce_wp_users
+algolia_autocomplete_debounce_wp_terms_category
+algolia_autocomplete_debounce_wp_terms_post_tag
+algolia_autocomplete_debounce_wp_terms_my_custom_taxonomy
+```
+Note that the Algolia Autocomplete settings must be saved after creating one of the above filters.
+
 ## 2.8.0
 * Added: Filter to customize Algolia SearchClient configuration with connect/read/write timeouts.
 * Updated: Prevent table content from being concatenated. Thanks @rodrigo-arias